Moves urlParams logic out of store and into context directory, updates references
