+++ x-pack/plugins/apm/public/components/app/ServiceOverview/index.ts
@@ -6,7 +6,7 @@
 
 import { connect } from 'react-redux';
 import { IReduxState } from '../../../store/rootReducer';
-import { getUrlParams } from '../../../store/urlParams';
+import { getUrlParams } from '../../../context/UrlParamsContext';
 import { ServiceOverview as View } from './view';
 
 function mapStateToProps(state = {} as IReduxState) {
