+++ x-pack/plugins/apm/public/components/app/TraceOverview/index.tsx
@@ -6,7 +6,7 @@
 
 import { connect } from 'react-redux';
 import { IReduxState } from '../../../store/rootReducer';
-import { getUrlParams } from '../../../store/urlParams';
+import { getUrlParams } from '../../../context/UrlParamsContext';
 import { TraceOverview as View } from './view';
 
 function mapStateToProps(state = {} as IReduxState) {
