+++ x-pack/plugins/apm/public/context/UrlParamsContext/constants.ts
@@ -1,0 +1,14 @@
+/*
+ * Copyright Elasticsearch B.V. and/or licensed to Elasticsearch B.V. under one
+ * or more contributor license agreements. Licensed under the Elastic License;
+ * you may not use this file except in compliance with the Elastic License.
+ */
+
+export const TIME_RANGE_REFRESH = 'TIME_RANGE_REFRESH';
+export const LOCATION_UPDATE = 'LOCATION_UPDATE';
+export const TIMEPICKER_DEFAULTS = {
+  rangeFrom: 'now-24h',
+  rangeTo: 'now',
+  refreshPaused: 'true',
+  refreshInterval: '0'
+};
