+++ x-pack/plugins/apm/public/context/UrlParamsContext/index.tsx
@@ -1,0 +1,84 @@
+/*
+ * Copyright Elasticsearch B.V. and/or licensed to Elasticsearch B.V. under one
+ * or more contributor license agreements. Licensed under the Elastic License;
+ * you may not use this file except in compliance with the Elastic License.
+ */
+
+import React, { createContext, useReducer } from 'react';
+import { Location } from 'history';
+import { useLocation } from '../../hooks/useLocation';
+import { IUrlParams } from './types';
+import { LOCATION_UPDATE, TIME_RANGE_REFRESH } from './constants';
+import { getParsedDate } from './helpers';
+import { resolveUrlParams } from './resolveUrlParams';
+
+// TODO: Remove this selector once all its uses are converted to the hook or context
+export function getUrlParams(state: any) {
+  return state.urlParams;
+}
+
+interface TimeRange {
+  rangeFrom: string;
+  rangeTo: string;
+}
+
+interface LocationAction {
+  type: typeof LOCATION_UPDATE;
+  location: Location;
+}
+
+interface TimeRangeRefreshAction {
+  type: typeof TIME_RANGE_REFRESH;
+  time: TimeRange;
+}
+
+type RefreshTimeRangeFunction = (time: TimeRange) => void;
+
+const defaultRefresh = (() => null) as RefreshTimeRangeFunction;
+
+export function urlParamsReducer(
+  state: IUrlParams = {},
+  action: LocationAction | TimeRangeRefreshAction
+): IUrlParams {
+  switch (action.type) {
+    case LOCATION_UPDATE: {
+      return resolveUrlParams(action.location);
+    }
+
+    case TIME_RANGE_REFRESH:
+      return {
+        ...state,
+        start: getParsedDate(action.time.rangeFrom),
+        end: getParsedDate(action.time.rangeTo)
+      };
+
+    default:
+      return state;
+  }
+}
+
+const UrlParamsContext = createContext({
+  urlParams: {} as IUrlParams,
+  refreshTimeRange: defaultRefresh
+});
+
+const UrlParamsProvider: React.FC<{}> = ({ children }) => {
+  const location = useLocation();
+  const [urlParams, dispatch] = useReducer(
+    urlParamsReducer,
+    resolveUrlParams(location)
+  );
+
+  function refreshTimeRange(time: TimeRange) {
+    dispatch({ type: TIME_RANGE_REFRESH, time });
+  }
+
+  return (
+    <UrlParamsContext.Provider
+      children={children}
+      value={{ urlParams, refreshTimeRange }}
+    />
+  );
+};
+
+export { UrlParamsContext, UrlParamsProvider };
