+++ x-pack/plugins/apm/public/context/UrlParamsContext/types.ts
@@ -1,0 +1,26 @@
+/*
+ * Copyright Elasticsearch B.V. and/or licensed to Elasticsearch B.V. under one
+ * or more contributor license agreements. Licensed under the Elastic License;
+ * you may not use this file except in compliance with the Elastic License.
+ */
+
+export interface IUrlParams {
+  detailTab?: string;
+  end?: string;
+  errorGroupId?: string;
+  flyoutDetailTab?: string;
+  kuery?: string;
+  rangeFrom?: string;
+  rangeTo?: string;
+  refreshInterval?: number;
+  refreshPaused?: boolean;
+  serviceName?: string;
+  sortDirection?: string;
+  sortField?: string;
+  start?: string;
+  traceId?: string;
+  transactionId?: string;
+  transactionName?: string;
+  transactionType?: string;
+  waterfallItemId?: string;
+}
