+++ x-pack/plugins/apm/public/hooks/useTransactionDetailsCharts.ts
@@ -7,7 +7,7 @@
 import { useMemo } from 'react';
 import { loadTransactionDetailsCharts } from '../services/rest/apm/transaction_groups';
 import { getTransactionCharts } from '../store/selectors/chartSelectors';
-import { IUrlParams } from '../store/urlParams';
+import { IUrlParams } from '../context/UrlParamsContext/types';
 import { useFetcher } from './useFetcher';
 
 export function useTransactionDetailsCharts(urlParams: IUrlParams) {
