+++ x-pack/plugins/apm/public/hooks/useWaterfall.ts
@@ -7,7 +7,7 @@
 import { useMemo } from 'react';
 import { getWaterfall } from '../components/app/TransactionDetails/Transaction/WaterfallContainer/Waterfall/waterfall_helpers/waterfall_helpers';
 import { loadTrace } from '../services/rest/apm/traces';
-import { IUrlParams } from '../store/urlParams';
+import { IUrlParams } from '../context/UrlParamsContext/types';
 import { useFetcher } from './useFetcher';
 
 const INITIAL_DATA = { trace: [], errorsPerTransaction: {} };
